Fix DAB status printing because of strings in PROGMEM
The strings for the DAB status display are stored in PROGMEM. When using regular sprintf(), the string is not retrieved from the PROGMEM area. sprintf_P has a special format specifier, %S, to be able to use strings in PROGMEM directly. Use that specifier to get status display working again.